requirements.txt
async-generator==1.10
attrs==23.1.0
boto3==1.26.163
botocore==1.29.163
cachetools==5.3.1
certifi==2023.5.7
cffi==1.16.0
charset-normalizer==3.1.0
cryptography==42.0.5
dnspython==2.3.0
et-xmlfile==1.1.0
exceptiongroup==1.1.1
filelock==3.17.0
fsspec==2024.12.0
google-api-core==2.11.1
google-api-python-client==2.92.0
google-auth==2.21.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
googleapis-common-protos==1.59.1
h11==0.14.0
httplib2==0.22.0
idna==3.4
isodate==0.6.1
Jinja2==3.1.5
jmespath==1.0.1
lxml==5.2.1
MarkupSafe==3.0.2
memory-profiler==0.61.0
more-itertools==10.2.0
mpmath==1.3.0
networkx==3.4.2
numpy==1.25.0
oauthlib==3.2.2
openpyxl==3.1.2
outcome==1.2.0
pandas==2.0.2
pillow==11.1.0
platformdirs==4.2.0
protobuf==4.23.3
psutil==6.1.1
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.22
PyJWT==2.8.0
pymongo==4.3.3
pyparsing==3.1.0
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
requests==2.31.0
requests-file==2.0.0
requests-oauthlib==1.3.1
requests-toolbelt==1.0.0
rsa==4.9
s3transfer==0.6.1
scipy==1.15.1
selenium==4.10.0
simple-salesforce==1.12.6
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
sympy==1.13.1
torch==2.6.0
tqdm==4.67.1
trio==0.22.0
trio-websocket==0.10.3
typing_extensions==4.11.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.16
wsproto==1.2.0
zeep==4.2.1
